feat(post-api): setting up schema for post API routes
