+++ drizzle.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './src/db/schema.ts',
+  schema: './src/db/schema/*',
   out: './src/db/migrations',
   dialect: 'postgresql',
   dbCredentials: {
